split nps logic into two files
